Decouple window deletion from window expiration

+++ quixstreams/state/rocksdb/windowed/transaction.py
@@ -11,6 +11,8 @@ from quixstreams.state.serialization import DumpsFunc, LoadsFunc, serialize
 from .metadata import (
     LATEST_EXPIRED_WINDOW_CF_NAME,
     LATEST_EXPIRED_WINDOW_TIMESTAMP_KEY,
+    LATEST_DELETED_WINDOW_CF_NAME,
+    LATEST_DELETED_WINDOW_TIMESTAMP_KEY,
     LATEST_TIMESTAMP_KEY,
     LATEST_TIMESTAMPS_CF_NAME,
 )
@@ -85,7 +87,7 @@ class WindowedRocksDBPartitionTransaction(PartitionTransaction):
         timestamp_ms: int,
         prefix: bytes,
         window_timestamp_ms: Optional[int] = None,
-    ):
+    ) -> None:
         if timestamp_ms < 0:
             raise ValueError("Timestamp cannot be negative")
         self._validate_duration(start_ms=start_ms, end_ms=end_ms)
@@ -105,7 +107,7 @@ class WindowedRocksDBPartitionTransaction(PartitionTransaction):
         self.delete(key=key, prefix=prefix)
 
     def expire_windows(
-        self, watermark: int, prefix: bytes
+        self, watermark: int, prefix: bytes, delete: bool = True
     ) -> list[tuple[tuple[int, int], Any]]:
         """
         Get all expired windows from RocksDB up to the specified `watermark` timestamp.
@@ -124,6 +126,7 @@ class WindowedRocksDBPartitionTransaction(PartitionTransaction):
 
         :param watermark: The timestamp up to which windows are considered expired, inclusive.
         :param prefix: The key prefix for filtering windows.
+        :param delete: If True, expired windows will be deleted.
         :return: A sorted list of tuples in the format `((start, end), value)`.
         """
         latest_timestamp = self.get_latest_timestamp(prefix=prefix)
@@ -150,9 +153,56 @@ class WindowedRocksDBPartitionTransaction(PartitionTransaction):
                 prefix=prefix, timestamp_ms=last_expired__gt
             )
             # Delete expired windows from the state
-            for (start, end), _ in expired_windows:
-                self.delete_window(start, end, prefix=prefix)
+            if delete:
+                for (start, end), _ in expired_windows:
+                    self.delete_window(start, end, prefix=prefix)
         return expired_windows
+
+    def delete_windows(self, watermark: int, prefix: bytes) -> None:
+        """
+        Delete windows from RocksDB up to the specified `watermark` timestamp.
+
+        This method removes all window entries that have a start time less than or equal to the given
+        `watermark`. It ensures that expired data is cleaned up efficiently without affecting
+        unexpired windows.
+
+        How it works:
+        - It retrieves the start time of the last deleted window for the given prefix from the
+        deletion index. This minimizes redundant scans over already deleted windows.
+        - It iterates over the windows starting from the last deleted timestamp up to the `watermark`.
+        - Each window within this range is deleted from the database.
+        - After deletion, it updates the deletion index with the start time of the latest window
+        that was deleted to keep track of progress.
+
+        :param watermark: The timestamp up to which windows should be deleted, inclusive.
+        :param prefix: The key prefix used to identify and filter relevant windows.
+        """
+        start_from = self.get(
+            key=LATEST_DELETED_WINDOW_TIMESTAMP_KEY,
+            prefix=prefix,
+            cf_name=LATEST_DELETED_WINDOW_CF_NAME,
+            default=-1,
+        )
+
+        windows = self.get_windows(
+            start_from_ms=start_from,
+            start_to_ms=watermark,
+            prefix=prefix,
+        )
+
+        last_deleted__gt = None
+        for (start, end), _ in windows:
+            last_deleted__gt = start
+            self.delete_window(start, end, prefix=prefix)
+
+        # Save the start of the latest deleted window to the deletion index
+        if last_deleted__gt:
+            self.set(
+                key=LATEST_DELETED_WINDOW_TIMESTAMP_KEY,
+                value=last_deleted__gt,
+                prefix=prefix,
+                cf_name=LATEST_DELETED_WINDOW_CF_NAME,
+            )
 
     def get_windows(
         self,
